Invert files order in diff
Before: Home version is "old", dotfile is "new"
Now: Dotfile is "old", home version is "new"

This order makes more sense when an external program has replaced a symlink with a modified regular file.

However, it makes less sense when the user has 2 machines with differing dotfiles, and installs them on the second machine for the first time. This situation is far less common.