Remove stale links before creating new ones, not after
This fixes some edge cases, mainly when replacing a file with a directory with the same name, or vice versa.

It might have a negative impact on performance when renaming a deeply nested file, since the empty parent directories will be deleted and created again, instead of reused. However, that case is quite rare and the impact is likely minimal.